# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GatePass SaaS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type CompanyPlan = (typeof CompanyPlan)[keyof typeof CompanyPlan];

export const CompanyPlan = {
  starter: "starter",
  growth: "growth",
  enterprise: "enterprise",
} as const;

export interface Company {
  id: string;
  name: string;
  slug: string;
  logoUrl?: string | null;
  plan: CompanyPlan;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CompanyWithStats = Company & {
  userCount: number;
  officeCount: number;
};

export type CreateCompanyInputPlan =
  (typeof CreateCompanyInputPlan)[keyof typeof CreateCompanyInputPlan];

export const CreateCompanyInputPlan = {
  starter: "starter",
  growth: "growth",
  enterprise: "enterprise",
} as const;

export interface CreateCompanyInput {
  name: string;
  slug: string;
  plan?: CreateCompanyInputPlan;
}

export type UpdateCompanyInputPlan =
  (typeof UpdateCompanyInputPlan)[keyof typeof UpdateCompanyInputPlan];

export const UpdateCompanyInputPlan = {
  starter: "starter",
  growth: "growth",
  enterprise: "enterprise",
} as const;

export interface UpdateCompanyInput {
  name?: string;
  plan?: UpdateCompanyInputPlan;
  isActive?: boolean;
  logoUrl?: string;
}

export interface Office {
  id: string;
  companyId: string;
  name: string;
  city: string;
  address?: string | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateOfficeInput {
  name: string;
  city: string;
  address?: string;
}

export interface UpdateOfficeInput {
  name?: string;
  city?: string;
  address?: string;
  isActive?: boolean;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  super_admin: "super_admin",
  admin: "admin",
  security: "security",
  viewer: "viewer",
} as const;

export interface UserProfile {
  id: string;
  clerkId: string;
  companyId?: string | null;
  officeId?: string | null;
  name: string;
  email: string;
  role: UserProfileRole;
  isActive: boolean;
  createdAt: string;
  company?: Company | null;
  office?: Office | null;
}

export interface UpdateMeInput {
  name?: string;
  officeId?: string;
}

export type UpdateUserInputRole =
  (typeof UpdateUserInputRole)[keyof typeof UpdateUserInputRole];

export const UpdateUserInputRole = {
  admin: "admin",
  security: "security",
  viewer: "viewer",
} as const;

export interface UpdateUserInput {
  role?: UpdateUserInputRole;
  officeId?: string;
  isActive?: boolean;
}

export type AdminUpdateUserInputRole =
  (typeof AdminUpdateUserInputRole)[keyof typeof AdminUpdateUserInputRole];

export const AdminUpdateUserInputRole = {
  super_admin: "super_admin",
  admin: "admin",
  security: "security",
  viewer: "viewer",
} as const;

export interface AdminUpdateUserInput {
  role?: AdminUpdateUserInputRole;
  companyId?: string;
  officeId?: string;
  isActive?: boolean;
}

export type VisitorStatus = (typeof VisitorStatus)[keyof typeof VisitorStatus];

export const VisitorStatus = {
  Pending: "Pending",
  Checked_In: "Checked In",
  On_Break: "On Break",
  Checked_Out: "Checked Out",
} as const;

export interface Visitor {
  id: string;
  companyId: string;
  officeId: string;
  visitorId: string;
  name: string;
  phone?: string | null;
  email?: string | null;
  company?: string | null;
  host: string;
  type: string;
  status: VisitorStatus;
  purpose?: string | null;
  idType?: string | null;
  idNumber?: string | null;
  photoUrl?: string | null;
  vehicleNumber?: string | null;
  checkInTime?: string | null;
  checkOutTime?: string | null;
  expectedCheckout?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface RegisterVisitorInput {
  officeId: string;
  name: string;
  phone?: string;
  email?: string;
  company?: string;
  host: string;
  type: string;
  purpose?: string;
  idType?: string;
  idNumber?: string;
  photoUrl?: string;
  vehicleNumber?: string;
  checkInNow?: boolean;
}

export interface PublicRegisterInput {
  companySlug: string;
  officeId: string;
  name: string;
  phone?: string;
  email?: string;
  company?: string;
  host: string;
  type: string;
  purpose?: string;
  idType?: string;
  idNumber?: string;
  photoUrl?: string;
  vehicleNumber?: string;
}

export interface VisitorLog {
  id: string;
  vid: string;
  visitorId: string;
  companyId: string;
  officeId: string;
  name: string;
  action: string;
  note?: string | null;
  ts: string;
}

export type GatePassStatus =
  (typeof GatePassStatus)[keyof typeof GatePassStatus];

export const GatePassStatus = {
  Open: "Open",
  Closed: "Closed",
} as const;

export type GatePassItemsItem = {
  name: string;
  qty: number;
  unit: string;
};

export interface GatePass {
  id: string;
  companyId: string;
  officeId: string;
  passId: string;
  purpose: string;
  type: string;
  status: GatePassStatus;
  vendorName?: string | null;
  vehicleNo?: string | null;
  driverName?: string | null;
  requestedBy: string;
  items: GatePassItemsItem[];
  itemCount: number;
  openedAt: string;
  closedAt?: string | null;
  notes?: string | null;
  createdAt: string;
}

export type CreateGatePassInputItemsItem = {
  name: string;
  qty: number;
  unit: string;
};

export interface CreateGatePassInput {
  officeId: string;
  purpose: string;
  type: string;
  vendorName?: string;
  vehicleNo?: string;
  driverName?: string;
  requestedBy: string;
  items: CreateGatePassInputItemsItem[];
  notes?: string;
}

export interface GpLog {
  id: string;
  passId: string;
  gatePassId: string;
  companyId: string;
  officeId: string;
  type: string;
  by: string;
  note?: string | null;
  ts: string;
}

/**
 * Not authenticated
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Not authorized
 */
export type ForbiddenResponse = ErrorResponse;

export type ListVisitorsParams = {
  officeId: string;
  /**
   * ISO date string (YYYY-MM-DD), defaults to today
   */
  date?: string;
  status?: string;
};

export type ListVisitorLogsParams = {
  officeId: string;
  date?: string;
};

export type ListGatePassesParams = {
  officeId: string;
  status?: string;
  date?: string;
};

export type ListGpLogsParams = {
  officeId: string;
  date?: string;
};

export type AdminListUsersParams = {
  companyId?: string;
};