# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GatePass SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offices:
    get:
      operationId: listOffices
      tags: [offices]
      summary: List offices for the current user's company
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of offices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Office"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createOffice
      tags: [offices]
      summary: Create a new office
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOfficeInput"
      responses:
        "201":
          description: Created office
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Office"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /offices/{officeId}:
    patch:
      operationId: updateOffice
      tags: [offices]
      summary: Update an office
      security:
        - clerkAuth: []
      parameters:
        - name: officeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOfficeInput"
      responses:
        "200":
          description: Updated office
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Office"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteOffice
      tags: [offices]
      summary: Delete (deactivate) an office
      security:
        - clerkAuth: []
      parameters:
        - name: officeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users in the current company (admin only)
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/{userId}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user's role or office
      security:
        - clerkAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Deactivate a user
      security:
        - clerkAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deactivated
...
  /gate-passes:
    get:
      operationId: listGatePasses
      tags: [gate-passes]
      summary: List gate passes
      security:
        - clerkAuth: []
      parameters:
        - name: officeId
          in: query
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: date
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of gate passes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GatePass"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createGatePass
      tags: [gate-passes]
      summary: Create a new gate pass
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGatePassInput"
      responses:
        "201":
          description: Created gate pass
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatePass"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /gate-passes/{gpId}/close:
    post:
      operationId: closeGatePass
      tags: [gate-passes]
      summary: Close a gate pass
      security:
        - clerkAuth: []
      parameters:
        - name: gpId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated gate pass
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatePass"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /admin/companies:
    get:
      operationId: adminListCompanies
      tags: [admin]
      summary: List all companies (super admin only)
      security:
        - clerkAuth: []
      responses:
        "200":
          description: All companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyWithStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: adminCreateCompany
      tags: [admin]
      summary: Create a company (super admin only)
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyInput"
      responses:
        "201":
          description: Created company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/companies/{companyId}:
    patch:
      operationId: adminUpdateCompany
      tags: [admin]
      summary: Update a company (super admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyInput"
      responses:
        "200":
          description: Updated company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      operationId: adminDeleteCompany
      tags: [admin]
      summary: Deactivate a company (super admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deactivated
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (super admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: companyId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{userId}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update any user (super admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Not authorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Company:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
          nullable: true
        plan:
          type: string
          enum: [starter, growth, enterprise]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, slug, plan, isActive, createdAt, updatedAt]
    CompanyWithStats:
      allOf:
        - $ref: "#/components/schemas/Company"
        - type: object
          properties:
            userCount:
              type: integer
            officeCount:
              type: integer
          required: [userCount, officeCount]
    CreateCompanyInput:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
          enum: [starter, growth, enterprise]
      required: [name, slug]
    UpdateCompanyInput:
      type: object
      properties:
        name:
          type: string
        plan:
          type: string
          enum: [starter, growth, enterprise]
        isActive:
          type: boolean
        logoUrl:
          type: string
    Office:
      type: object
      properties:
        id:
          type: string
        companyId:
          type: string
        name:
          type: string
        city:
          type: string
        address:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, companyId, name, city, isActive, createdAt, updatedAt]
    CreateOfficeInput:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        address:
          type: string
      required: [name, city]
    UpdateOfficeInput:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        address:
          type: string
        isActive:
          type: boolean
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        companyId:
          type: string
          nullable: true
        officeId:
          type: string
          nullable: true
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, admin, security, viewer]
        isActive:
          type: boolean
        createdAt:
          type: string
        company:
          $ref: "#/components/schemas/Company"
          nullable: true
        office:
          $ref: "#/components/schemas/Office"
          nullable: true
      required: [id, clerkId, name, email, role, isActive, createdAt]
...
    UpdateUserInput:
      type: object
      properties:
        role:
          type: string
          enum: [admin, security, viewer]
        officeId:
          type: string
        isActive:
          type: boolean
    AdminUpdateUserInput:
      type: object
      properties:
        role:
          type: string
          enum: [super_admin, admin, security, viewer]
        companyId:
          type: string
        officeId:
          type: string
        isActive:
          type: boolean
...
    GatePass:
      type: object
      properties:
        id:
          type: string
        companyId:
          type: string
        officeId:
          type: string
        passId:
          type: string
        purpose:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [Open, Closed]
        vendorName:
          type: string
          nullable: true
        vehicleNo:
          type: string
          nullable: true
        driverName:
          type: string
          nullable: true
        requestedBy:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              qty:
                type: integer
              unit:
                type: string
            required: [name, qty, unit]
        itemCount:
          type: integer
        openedAt:
          type: string
        closedAt:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, companyId, officeId, passId, purpose, type, status, requestedBy, items, itemCount, openedAt, createdAt]
    CreateGatePassInput:
      type: object
      properties:
        officeId:
          type: string
        purpose:
          type: string
        type:
          type: string
        vendorName:
          type: string
        vehicleNo:
          type: string
        driverName:
          type: string
        requestedBy:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              qty:
                type: integer
              unit:
                type: string
            required: [name, qty, unit]
        notes:
          type: string
      required: [officeId, purpose, type, requestedBy, items]
...