# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GatePass SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { CompanyPlan } from "./companyPlan";

export interface Company {
  id: string;
  name: string;
  slug: string;
  logoUrl?: string | null;
  plan: CompanyPlan;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}