# lib/api-zod/src/generated/types/companyWithStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GatePass SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { Company } from "./company";

export type CompanyWithStats = Company & {
  userCount: number;
  officeCount: number;
};