# lib/api-zod/src/generated/types/office.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GatePass SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface Office {
  id: string;
  companyId: string;
  name: string;
  city: string;
  address?: string | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}