# lib/api-zod/src/generated/types/updateOfficeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GatePass SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateOfficeInput {
  name?: string;
  city?: string;
  address?: string;
  isActive?: boolean;
}